Move simulated sample stage class to its own module
